refactor(backend): migrate routes into proper modules (HD-4)
Three route files previously lived in app/routes/ outside the
modules/ architecture. Migrated them into proper PascalCase module
directories:

- citation.routes.ts → modules/CitationExport/citationExport.routes.ts
- discussion.routes.ts → modules/Discussion/discussion.routes.ts
- activityLog.routes.ts → modules/ActivityLog/activityLog.routes.ts

Updated all relative imports (../services/ → ../../services/, etc.)
and routes/index.ts to import from new locations.

Removed stale legacy route comments.

Phase 8.2.D: Architecture Normalization
